+++ shm/include/game.hpp
@@ -2,12 +2,12 @@
 
 #include <memory>
 
-#include "PrintCargo.hpp"
-#include "Travel.hpp"
+#include "sell.hpp"
+#include "travel.hpp"
 #include "buy.hpp"
-#include "Sell.hpp"
 #include "map.hpp"
 #include "player.hpp"
+#include "print_cargo.hpp"
 #include "ship.hpp"
 #include "time.hpp"
