+++ shm/src/game.cpp
@@ -1,6 +1,5 @@
 #include <iostream>
 
-#include "Travel.hpp"
 #include "game.hpp"
 #include "locators.hpp"
 #include "time.hpp"
