+++ shm/src/print_cargo.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 
 #include "player.hpp"
-#include "PrintCargo.hpp"
+#include "print_cargo.hpp"
 
 void PrintCargo::execute(Player* player) {
     size_t cargoIndex = 0;
